requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
Pillow==8.3.1
python-dotenv==0.19.0
rembg==2.0.30
numpy==1.21.0
opencv-python==4.5.3.56
torch==1.9.0
torchvision==0.10.0 